Using floats when calculating approximation of filters. Removed taylormath. Fixed changing white channel on animations. Updated source control. Updated statistics in Readme. Added namespace for ledc

+++ main/src/common_types.h
@@ -4,7 +4,7 @@
 #include <array>
 #include "lib/fixedpoint/fixedpoint.h"
 
-using fixed32_c = fixedpoint<std::int32_t, std::int32_t, 16>;
+using fixed32_c = fixedpoint<std::int32_t, std::int64_t, 16>;
 using fixed32_f = fixedpoint<std::int64_t, std::int64_t, 24>;
 using ColorChannels = std::array<fixed32_c, 4>;
